+++ packages/studio-plugin/tests/parsers/TypeNodeParsingHelper.test.ts
@@ -2,7 +2,7 @@ import { SyntaxKind } from "ts-morph";
 import { PropValueType } from "../../lib";
 import TypeNodeParsingHelper, {
   ParsedTypeKind,
-} from "../../src/parsers/helpers/ShapeParsingHelper";
+} from "../../src/parsers/helpers/TypeNodeParsingHelper";
 import createTestSourceFile from "../__utils__/createTestSourceFile";
 
 const externalShapeParser = jest.fn();
